Moved core class into Core
